+++ sync-for-commerce/src/sdk/models/operations/getsyncbyid.ts
@@ -7,6 +7,9 @@ import * as shared from "../shared";
 import { AxiosResponse } from "axios";
 
 export class GetSyncByIdRequest extends SpeakeasyBase {
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
@@ -22,8 +25,11 @@ export class GetSyncByIdResponse extends SpeakeasyBase {
      * Success
      */
     @SpeakeasyMetadata()
-    companySyncStatus?: shared.CompanySyncStatus;
+    companySyncStatus?: Record<string, any>;
 
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -33,9 +39,15 @@ export class GetSyncByIdResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 }