+++ sync-for-commerce/src/sdk/models/operations/getsyncstatus.ts
@@ -3,10 +3,12 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
-import * as shared from "../shared";
 import { AxiosResponse } from "axios";
 
 export class GetSyncStatusRequest extends SpeakeasyBase {
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 }
@@ -18,6 +20,9 @@ export class GetSyncStatusResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     badRequest?: any;
 
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -27,9 +32,15 @@ export class GetSyncStatusResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     notFound?: any;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 
@@ -37,5 +48,5 @@ export class GetSyncStatusResponse extends SpeakeasyBase {
      * Success
      */
     @SpeakeasyMetadata()
-    syncStatus?: shared.SyncStatus;
+    syncStatus?: Record<string, any>;
 }