+++ sync-for-commerce/src/sdk/models/operations/requestsync.ts
@@ -10,17 +10,29 @@ export class RequestSyncRequest extends SpeakeasyBase {
     @SpeakeasyMetadata({ data: "request, media_type=application/json" })
     syncToLatestArgs?: shared.SyncToLatestArgs;
 
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 }
 
 export class RequestSyncResponse extends SpeakeasyBase {
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 
@@ -28,5 +40,5 @@ export class RequestSyncResponse extends SpeakeasyBase {
      * Success
      */
     @SpeakeasyMetadata()
-    syncSummary?: shared.SyncSummary;
+    syncSummary?: Record<string, any>;
 }