+++ sync-for-commerce/src/sdk/sdk.ts
@@ -30,6 +30,7 @@ export type SDKProps = {
      * The security details required to authenticate the SDK
      */
     security?: shared.Security | (() => Promise<shared.Security>);
+
     /**
      * Allows overriding the default axios client used by the SDK
      */
@@ -57,8 +58,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "1.1";
-    sdkVersion = "2.1.0";
-    genVersion = "2.116.0";
+    sdkVersion = "2.2.0";
+    genVersion = "2.150.0";
+    userAgent = "speakeasy-sdk/typescript 2.2.0 2.150.0 1.1 @codat/sync-for-commerce";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);